[UIK-5008][dot] rewrite component to TS
