+++ semcore/dot/package.json
@@ -9,17 +9,12 @@
   "author": "UI-kit team <ui-kit-team@semrush.com>",
   "license": "MIT",
   "scripts": {
-    "build": "pnpm semcore-builder --source=js && pnpm vite build"
+    "build": "pnpm semcore-builder && pnpm vite build"
   },
   "exports": {
     "require": "./lib/cjs/index.js",
     "import": "./lib/esm/index.mjs",
     "types": "./lib/types/index.d.ts"
-  },
-  "dependencies": {
-    "@semcore/animation": "16.0.7",
-    "@semcore/flex-box": "16.0.11",
-    "@semcore/portal": "16.0.11"
   },
   "peerDependencies": {
     "@semcore/base-components": "^16.0.0"
