+++ semcore/dot/src/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Dot';
+export * from './Dot.type';
